+++ src/MarkdownEditorPackage.cs
@@ -10,22 +10,28 @@ namespace MarkdownEditor
     [Guid(PackageGuids.guidPackageString)]
     [PackageRegistration(UseManagedResourcesOnly = true, AllowsBackgroundLoading = true)]
     [InstalledProductRegistration("#110", "#112", Vsix.Version, IconResourceID = 400)]
-    [ProvideService(typeof(MarkdownLanguage), ServiceName = MarkdownLanguage.LanguageName, IsAsyncQueryable = true)]
+
     [ProvideLanguageService(typeof(MarkdownLanguage), MarkdownLanguage.LanguageName, 100, ShowDropDownOptions = true, DefaultToInsertSpaces = true, EnableCommenting = true, AutoOutlining = true)]
     [ProvideLanguageEditorOptionPage(typeof(Options), MarkdownLanguage.LanguageName, null, "Advanced", "#101", new[] { "markdown", "md" })]
-    [ProvideLanguageExtension(typeof(MarkdownLanguage), ".markdown")]
-    [ProvideLanguageExtension(typeof(MarkdownLanguage), ".md")]
-    [ProvideLanguageExtension(typeof(MarkdownLanguage), ".mdown")]
-    [ProvideLanguageExtension(typeof(MarkdownLanguage), ".mdwn")]
-    [ProvideLanguageExtension(typeof(MarkdownLanguage), ".mkd")]
-    [ProvideLanguageExtension(typeof(MarkdownLanguage), ".mkdn")]
-    [ProvideLanguageExtension(typeof(MarkdownLanguage), ".mmd")]
-    [ProvideLanguageExtension(typeof(MarkdownLanguage), ".rst")]
+
+    [ProvideEditorFactory(typeof(EditorFactory), 110)]
+    [ProvideEditorLogicalView(typeof(EditorFactory), VSConstants.LOGVIEWID.TextView_string, IsTrusted = true)]
+
+    [ProvideEditorExtension(typeof(EditorFactory), ".markdown", 0x32, NameResourceID = 110)]
+    [ProvideEditorExtension(typeof(EditorFactory), ".md", 0x32, NameResourceID = 110)]
+    [ProvideEditorExtension(typeof(EditorFactory), ".mdown", 0x32, NameResourceID = 110)]
+    [ProvideEditorExtension(typeof(EditorFactory), ".mdwn", 0x32, NameResourceID = 110)]
+    [ProvideEditorExtension(typeof(EditorFactory), ".mkd", 0x32, NameResourceID = 110)]
+    [ProvideEditorExtension(typeof(EditorFactory), ".mkdn", 0x32, NameResourceID = 110)]
+    [ProvideEditorExtension(typeof(EditorFactory), ".mmd", 0x32, NameResourceID = 110)]
+    [ProvideEditorExtension(typeof(EditorFactory), ".rst", 0x32, NameResourceID = 110)]
+    [ProvideEditorExtension(typeof(EditorFactory), ".*", 1, NameResourceID = 110)]
     [ProvideMenuResource("Menus.ctmenu", 1)]
     public sealed class MarkdownEditorPackage : Package
     {
         private static Options _options;
         private static object _syncRoot = new object();
+        private EditorFactory editorFactory;
 
         public static Options Options
         {
@@ -61,6 +67,9 @@ namespace MarkdownEditor
             var serviceContainer = this as IServiceContainer;
             var langService = new MarkdownLanguage(this);
             serviceContainer.AddService(typeof(MarkdownLanguage), langService, true);
+
+            editorFactory = new EditorFactory(this, langService.GetLanguageServiceGuid());
+            RegisterEditorFactory(editorFactory);
 
             base.Initialize();
         }